Update Id's to be long and update jwt key

+++ drift/src/main/java/drift/com/drift/api/APIConversationAPIBuilder.kt
@@ -21,7 +21,7 @@ import retrofit2.http.QueryMap
 internal interface APIConversationAPIBuilder {
 
     @GET("conversations/{conversationId}/messages")
-    fun getMessages(@Path("conversationId") conversationId: Int?): Call<ArrayList<Message>>
+    fun getMessages(@Path("conversationId") conversationId: Long?): Call<ArrayList<Message>>
 
     @GET("attachments")
     fun getAttachments(@Query("id") attachmentIds: List<Int>, @QueryMap imgixOptions: Map<String, String>): Call<ArrayList<Attachment>>
@@ -31,15 +31,15 @@ internal interface APIConversationAPIBuilder {
     fun getConversationsForEndUser(@Path("endUserId") endUserId: Long?): Call<ArrayList<ConversationExtra>>
 
     @POST("conversations/{conversationId}/messages")
-    fun postMessage(@Path("conversationId") conversationId: Int?, @Body messageRequest: MessageRequest): Call<Message>
+    fun postMessage(@Path("conversationId") conversationId: Long?, @Body messageRequest: MessageRequest): Call<Message>
 
     @POST("messages")
     fun createConversation(@Body payload: HashMap<String, Any>): Call<Message>
 
     @POST("https://conversation2.api.drift.com/messages/{messageId}/read")
-    fun markMessageAsRead(@Path("messageId") messageId: Int?): Call<ResponseBody>
+    fun markMessageAsRead(@Path("messageId") messageId: Long?): Call<ResponseBody>
 
     @POST("https://conversation2.api.drift.com/messages/{messageId}/read-until")
-    fun markMessageAsReadAlongWithPreviousMessages(@Path("messageId") messageId: Int?): Call<ResponseBody>
+    fun markMessageAsReadAlongWithPreviousMessages(@Path("messageId") messageId: Long?): Call<ResponseBody>
 
 }
